requirements.txt
streamlit==1.36.0
langchain
langchain-community
langchain-openai
langchain-google-genai
python-dotenv==1.0.1
pydantic==2.8.2
metapub==0.5.12
chromadb==0.4.24
google-generativeai